New `deconflict`
