+++ src/lem/path.rs
@@ -1,8 +1,6 @@
-use anyhow::Result;
-use indexmap::IndexMap;
 use std::collections::HashSet;
 
-use super::{symbol::Symbol, tag::Tag, var_map::VarMap, Block, Ctrl, Func, Op, Var};
+use super::{symbol::Symbol, tag::Tag, Block, Ctrl, Func, Op};
 
 #[derive(Clone, PartialEq, Eq, Hash)]
 pub(crate) enum PathNode {
@@ -80,94 +78,20 @@ impl Path {
     }
 }
 
-#[inline]
-fn insert_one(map: &mut VarMap<Var>, path: &Path, ptr: &Var) -> Var {
-    let new_ptr = Var(format!("{}.{}", path, ptr.name()).into());
-    map.insert(ptr.clone(), new_ptr.clone());
-    new_ptr
-}
+impl Func {
+    /// Computes the number of possible paths in a `Func` can take
+    pub fn num_paths(&self) -> usize {
+        self.block.num_paths()
+    }
 
-fn insert_many(map: &mut VarMap<Var>, path: &Path, ptrs: &[Var]) -> Vec<Var> {
-    ptrs.iter().map(|ptr| insert_one(map, path, ptr)).collect()
+    /// Asserts that all paths were visited by a set of frames. This is mostly
+    /// for testing purposes.
+    pub fn assert_all_paths_taken(&self, paths: &[Path]) {
+        assert_eq!(Path::num_paths_taken(paths), self.num_paths());
+    }
 }
 
 impl Block {
-    fn deconflict(
-        self,
-        path: &Path,
-        // `map` keeps track of the updated names of variables
-        map: &mut VarMap<Var>, // name -> path.name
-    ) -> Result<Self> {
-        let mut ops = Vec::with_capacity(self.ops.len());
-        for op in self.ops {
-            match op {
-                Op::Call(out, func, inp) => {
-                    let out = map.get_many_cloned(&out)?;
-                    let inp = map.get_many_cloned(&inp)?;
-                    ops.push(Op::Call(out, func, inp))
-                }
-                Op::Null(ptr, tag) => ops.push(Op::Null(insert_one(map, path, &ptr), tag)),
-                Op::Hash2(img, tag, preimg) => {
-                    let preimg = map.get_many_cloned(&preimg)?.try_into().unwrap();
-                    let img = insert_one(map, path, &img);
-                    ops.push(Op::Hash2(img, tag, preimg))
-                }
-                Op::Hash3(img, tag, preimg) => {
-                    let preimg = map.get_many_cloned(&preimg)?.try_into().unwrap();
-                    let img = insert_one(map, path, &img);
-                    ops.push(Op::Hash3(img, tag, preimg))
-                }
-                Op::Hash4(img, tag, preimg) => {
-                    let preimg = map.get_many_cloned(&preimg)?.try_into().unwrap();
-                    let img = insert_one(map, path, &img);
-                    ops.push(Op::Hash4(img, tag, preimg))
-                }
-                Op::Unhash2(preimg, img) => {
-                    let img = map.get_cloned(&img)?;
-                    let preimg = insert_many(map, path, &preimg);
-                    ops.push(Op::Unhash2(preimg.try_into().unwrap(), img))
-                }
-                Op::Unhash3(preimg, img) => {
-                    let img = map.get_cloned(&img)?;
-                    let preimg = insert_many(map, path, &preimg);
-                    ops.push(Op::Unhash3(preimg.try_into().unwrap(), img))
-                }
-                Op::Unhash4(preimg, img) => {
-                    let img = map.get_cloned(&img)?;
-                    let preimg = insert_many(map, path, &preimg);
-                    ops.push(Op::Unhash4(preimg.try_into().unwrap(), img))
-                }
-                Op::Hide(..) => todo!(),
-                Op::Open(..) => todo!(),
-            }
-        }
-        let ctrl = match self.ctrl {
-            Ctrl::MatchTag(var, cases) => {
-                let mut new_cases = Vec::with_capacity(cases.len());
-                for (tag, case) in cases {
-                    let new_case = case.deconflict(&path.push_tag(&tag), &mut map.clone())?;
-                    new_cases.push((tag, new_case));
-                }
-                Ctrl::MatchTag(map.get_cloned(&var)?, IndexMap::from_iter(new_cases))
-            }
-            Ctrl::MatchSymbol(var, cases, def) => {
-                let mut new_cases = Vec::with_capacity(cases.len());
-                for (symbol, case) in cases {
-                    let new_case = case.deconflict(&path.push_symbol(&symbol), &mut map.clone())?;
-                    new_cases.push((symbol.clone(), new_case));
-                }
-                Ctrl::MatchSymbol(
-                    map.get_cloned(&var)?,
-                    IndexMap::from_iter(new_cases),
-                    Box::new(def.deconflict(&path.push_default(), map)?),
-                )
-            }
-            Ctrl::Return(o) => Ctrl::Return(map.get_many_cloned(&o)?),
-        };
-        Ok(Block { ops, ctrl })
-    }
-
-    /// Computes the number of possible paths in a `Block`
     fn num_paths(&self) -> usize {
         let mut num_paths = 1;
         for op in &self.ops {
@@ -185,37 +109,5 @@ impl Block {
             Ctrl::Return(..) => 1,
         };
         num_paths
-    }
-}
-
-impl Func {
-    /// Removes conflicting names in parallel logical LEM paths. While these
-    /// conflicting names shouldn't be an issue for interpretation, they are
-    /// problematic when we want to generate the constraints for the LEM, since
-    /// conflicting names would cause different allocations to be bound the same
-    /// name.
-    ///
-    /// The conflict resolution is achieved by changing variables so that
-    /// their names are prepended by the paths where they're declared.
-    ///
-    /// Note: this function is not supposed to be called manually. It's used by
-    /// `LEM::new`, which is the API that should be used directly.
-    pub(crate) fn deconflict(mut self) -> Result<Self> {
-        let mut map = VarMap::new();
-        for i in self.input_params.iter() {
-            map.insert(i.clone(), i.clone());
-        }
-        self.block = self.block.deconflict(&Path::default(), &mut map)?;
-        Ok(self)
-    }
-
-    pub fn num_paths(&self) -> usize {
-        self.block.num_paths()
-    }
-
-    /// Asserts that all paths were visited by a set of frames. This is mostly
-    /// for testing purposes.
-    pub fn assert_all_paths_taken(&self, paths: &[Path]) {
-        assert_eq!(Path::num_paths_taken(paths), self.num_paths());
     }
 }
